check jq dependency

+++ shell/tutorial.sh
@@ -28,40 +28,45 @@
 # }
 # ew0KCSJuYW1lIjoiVHV0b3JpYWwgQXBwIiwNCgkiaWRlbnRpZmllciI6ImV1Lm93bnlvdXJkYXRhLnR1dG9yaWFsIiwNCgkidHlwZSI6ImV4dGVybmFsIiwNCgkiZGVzY3JpcHRpb24iOiJtYW5pZmVzdCBmb3IgdHV0b3JpYWwgYXBwIiwNCgkicGVybWlzc2lvbnMiOlsiZXUub3dueW91cmRhdGEuKjpyZWFkIiwNCgkJImV1Lm93bnlvdXJkYXRhLio6d3JpdGUiLA0KCQkiZXUub3dueW91cmRhdGEuKjp1cGRhdGUiLA0KCQkiZXUub3dueW91cmRhdGEuKjpkZWxldGUiXQ0KfQ==
 
-export PIA_URL=http://localhost:8080
-export APP_KEY=eu.ownyourdata.tutorial
-export APP_SECRET=l1InCBVMY0o31CJDcJKo # !!!update
+CONTINUE=true
+command -v jq >/dev/null 2>&1 || { echo >&2 "Fehler: JSON processor 'jq' ist nicht installiert."; CONTINUE=false;}
 
-# get token
-export TOKEN=`curl $APP_KEY:$APP_SECRET@${PIA_URL#*//}/oauth/token -d grant_type=client_credentials | jq -r '.access_token'`
-value=$(<tmp.txt)
+if $CONTINUE; then
+	export PIA_URL=http://localhost:8080
+	export APP_KEY=eu.ownyourdata.tutorial
+	export APP_SECRET=l1InCBVMY0o31CJDcJKo # !!!update
 
-# write data - simple JSON
-export CONTENT="{\"value\": 1}"
-curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/$APP_KEY/items
-# wildcard testing
-# OK
-# curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/eu.ownyourdata.sample/items
-# Not OK
-# curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/eu.sellyourdata.biz/items
+	# get token
+	export TOKEN=`curl -s $APP_KEY:$APP_SECRET@${PIA_URL#*//}/oauth/token -d grant_type=client_credentials | jq -r '.access_token'`
+	value=$(<tmp.txt)
 
-# write data - large dataset
-# export CONTENT=`head -c 99988 < /dev/zero | tr '\0' '\141'`
-# echo $CONTENT > tmp.txt
-# value=$(<tmp.txt)
-# curl -X POST -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -d @- "$PIA_URL/api/repos/$APP_KEY/items" <<tmp.txt
-# { "value": "$value"}
-# tmp.txt
+	# write data - simple JSON
+	export CONTENT="{\"value\": 1}"
+	curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/$APP_KEY/items
+	# wildcard testing
+	# OK
+	# curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/eu.ownyourdata.sample/items
+	# Not OK
+	# curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X POST -d "$CONTENT" $PIA_URL/api/repos/eu.sellyourdata.biz/items
 
-# update data
-export CONTENT='{"id": "1", "value": 0}'
-curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X PUT -d "$CONTENT" $PIA_URL/api/repos/$APP_KEY/items
+	# write data - large dataset
+	# export CONTENT=`head -c 99988 < /dev/zero | tr '\0' '\141'`
+	# echo $CONTENT > tmp.txt
+	# value=$(<tmp.txt)
+	# curl -X POST -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -d @- "$PIA_URL/api/repos/$APP_KEY/items" <<tmp.txt
+	# { "value": "$value"}
+	# tmp.txt
 
-# read data
-curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X GET $PIA_URL/api/repos/$APP_KEY/items
+	# update data
+	export CONTENT='{"id": "1", "value": 0}'
+	curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X PUT -d "$CONTENT" $PIA_URL/api/repos/$APP_KEY/items
 
-# delete data
-curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X DELETE $PIA_URL/api/repos/$APP_KEY/items/1
+	# read data
+	curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X GET $PIA_URL/api/repos/$APP_KEY/items
 
-# read data
-curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X GET $PIA_URL/api/repos/$APP_KEY/items
+	# delete data
+	curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X DELETE $PIA_URL/api/repos/$APP_KEY/items/1
+
+	# read data
+	curl -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" -X GET $PIA_URL/api/repos/$APP_KEY/items
+fi
